refactor(utils): improve separation of concerns and reduce abstraction cost
- Split xmlUtils.ts (599 LOC) into focused modules:
  - xmlCdata.ts: CDATA section handling
  - xmlFormatDetection.ts: HTML/LaTeX/Markdown detection
  - xmlConversion.ts: Pandoc/Turndown format conversion
  - xmlExtraction.ts: XML content extraction utilities
  - xmlUtils.ts: Barrel export for backward compatibility

- Remove unnecessary abstraction:
  - Delete baseFileUtils.ts wrapper (was just re-exporting from common)
  - Update files/index.ts to re-export directly from common module

- Use centralized Zod schemas to reduce duplication:
  - UsageTypes.ts: Use StreamTabIdSchema and StorageKeySchema
  - ProgressViewMessageHandler.ts: Use StreamTabIdSchema

- Add path utilities to reduce scattered duplication:
  - getExtensionLowercase(): Consolidates .toLowerCase().endsWith() pattern
  - hasExtension(): Case-insensitive extension checking
  - getFilenameWithoutExtension(): Consolidates basename(f, extname(f))

- Remove deprecated code:
  - Remove outputFormat alias (use OutputFormat enum directly)

- Fix pre-existing TypeScript error:
  - BaseReflectionAgent.ts: Remove invalid type argument from createReflectionRunFlow
