--- src/utils/files/baseFileUtils.ts
@@ -1,1 +1,0 @@
-export { getEffectiveBaseFile } from '@common/modules/files/baseFileUtils.js';
